@@ -43,17 +43,21 @@ classifiers =
      Natural Language :: English
      Framework :: Django
      Framework :: Django :: 2.2
-     Programming Language :: Python :: 3.6
+     Framework :: Django :: 3.0
+     Framework :: Django :: 3.2
      Programming Language :: Python :: 3.7
+     Programming Language :: Python :: 3.8
+     Programming Language :: Python :: 3.9
+     Programming Language :: Python :: 3.10
 
 [options]
 include_package_data = True
 install_requires =
-    django-cms>=3.6
+    django-cms>=3.7
 setup_requires =
     setuptools
 packages = djangocms_page_sitemap
-python_requires = >=3.6
+python_requires = >=3.7
 zip_safe = False
 test_suite = cms_helper.run
 
@@ -63,7 +67,7 @@ djangcms_page_sitemap = *.html *.png *.gif *js *jpg *jpeg *svg *py *mo *po
 
 [options.extras_require]
 docs =
-    django<3.1
+    django<4.0
 
 [upload]
 repository = https://upload.pypi.org/legacy/
