@@ -1,8 +1,8 @@
-from django.conf.urls import url
 from django.contrib.sitemaps.views import sitemap
+from django.urls import re_path
 
 from .sitemap import ExtendedSitemap
 
 urlpatterns = [
-    url(r"^sitemap\.xml$", sitemap, {"sitemaps": {"cmspages": ExtendedSitemap}}),
+    re_path(r"^sitemap\.xml$", sitemap, {"sitemaps": {"cmspages": ExtendedSitemap}}),
 ]
