@@ -1,9 +1,10 @@
 from cms.utils.conf import get_cms_setting
 from django.conf import settings
-from django.conf.urls import include, url
+from django.conf.urls import include
 from django.conf.urls.i18n import i18n_patterns
 from django.contrib import admin
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
+from django.urls import re_path
 from django.views.static import serve
 
 from djangocms_page_sitemap import sitemap_urls
@@ -11,22 +12,22 @@ from djangocms_page_sitemap import sitemap_urls
 admin.autodiscover()
 
 urlpatterns = [
-    url(
+    re_path(
         r"^media/(?P<path>.*)$",
         serve,
         {"document_root": settings.MEDIA_ROOT, "show_indexes": True},
     ),
-    url(
+    re_path(
         r"^media/cms/(?P<path>.*)$",
         serve,
         {"document_root": get_cms_setting("MEDIA_ROOT"), "show_indexes": True},
     ),
-    url(r"^", include(sitemap_urls)),
+    re_path(r"^", include(sitemap_urls)),
 ]
 
 urlpatterns += staticfiles_urlpatterns()
 
 urlpatterns += i18n_patterns(
-    url(r"^admin/", admin.site.urls),
-    url(r"^", include("cms.urls")),
+    re_path(r"^admin/", admin.site.urls),
+    re_path(r"^", include("cms.urls")),
 )
